docker-swarm/microservices-demo/worker/app/requirements.txt
requests==2.31.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-instrumentation-urllib==0.51b0
opentelemetry-exporter-otlp==1.30.0